+++ R/fileMDB.R
@@ -829,20 +829,11 @@ data_file_size <- function(x, hr=FALSE){
    stopifnot(
       length(i)==1
    )
-   ## Rstudio hack to avoid DB call when just looking for names
-   cc <- grep('.rs.getCompletionsDollar', deparse(sys.calls()), value=FALSE)
-   if(length(cc)!=0){
-      invisible(NULL)
+   ## Rstudio hack to avoid file call when just looking for names
+   if(.is_called_by_rs_function()){
+      return(.get_virtual_tables(x, i)[[1]])
    }else{
-      cc <- c(
-         grep('.rs.valueContents', deparse(sys.calls()), value=FALSE),
-         grep('.rs.explorer.inspectObject', deparse(sys.calls()), value=FALSE)
-      )
-      if(length(cc)!=0){
-         invisible(as.character(data_files(x)$dataFiles[i]))
-      }else{
-         return(data_tables(x, dplyr::all_of(i))[[1]])
-      }
+      return(data_tables(x, dplyr::all_of(i))[[1]])
    }
 }
 #' @rdname fileMDB
@@ -862,10 +853,9 @@ data_file_size <- function(x, hr=FALSE){
 #' @export
 #'
 as.list.fileMDB <- function(x, ...){
-   ## Rstudio hack to avoid DB call when just looking for names
-   cc <- grep('.rs.getCompletionsDollar', deparse(sys.calls()), value=FALSE)
-   if(length(cc)!=0){
-      invisible(NULL)
+   ## Rstudio hack to avoid file call when just looking for names
+   if(.is_called_by_rs_function()){
+      return(.get_virtual_tables(x, ...))
    }else{
       return(data_tables(x, ...))
    }