+++ R/memoMDB.R
@@ -514,22 +514,12 @@ dims.memoMDB <- function(x, ...){
    stopifnot(
       length(i)==1
    )
-   ## Rstudio hack to avoid DB call when just looking for names
-   cc <- grep('.rs.getCompletionsDollar', deparse(sys.calls()), value=FALSE)
-   if(length(cc)!=0){
-      invisible(NULL)
+   ## Rstudio hack to avoid call when just looking for names
+   if(.is_called_by_rs_function()){
+      return(.get_virtual_tables(x, i)[[1]])
    }else{
-      cc <- c(
-         grep('.rs.valueContents', deparse(sys.calls()), value=FALSE),
-         grep('.rs.explorer.inspectObject', deparse(sys.calls()), value=FALSE)
-      )
-      if(length(cc)!=0){
-         invisible()
-      }else{
-         return(data_tables(x, dplyr::all_of(i))[[1]])
-      }
+      return(data_tables(x, dplyr::all_of(i))[[1]])
    }
-   # return(data_tables(x, dplyr::all_of(i))[[1]])
 }
 #' @rdname memoMDB
 #'